requirements.txt
mypy>=0.76
flake8>=3.7.9
pylint>=2.4.4
pytest>=5.3.5
requests>=2.21
pillow>=7.0
jsonschema>=3.0.1
beautifulsoup4>=4.8.2
tinydb>=3.15.2
pydantic<2.0
opencv-python>=4.13
scikit-image>=0.26
ddddocr>=1.6
numpy>=2.4
